requirements.txt
numpy
discord.py
python-dotenv